Remove modlog ignore for voice verif
When a user leaves and rejoins, and then asks for the voice verification role back in modmail, we want a way to see if they have every been given the role.

By removing this modlog ignore, it allows us to check the user-log channel for user updates to see if they were ever assigned the role.